Algumas modificacoes no robo e reorganizacao do arquivo em bibliotecas
