--- Trabalho 2/makefile
@@ -1,7 +1,0 @@
-CC = g++
-CFLAGS = -Wall -lGL -lglut -lGLU
-TARGET = robo
-all: $(TARGET)
-$(TARGET): $(TARGET).cpp
-	$(CC) $(TARGET).cpp $(CFLAGS) -o $(TARGET)
-clean: $(RM) $(TARGET)
